clean silicon according to seismic specs
